+++ src/Peridigm_PdQuickGridDiscretization.cpp
@@ -83,7 +83,7 @@ PeridigmNS::PdQuickGridDiscretization::PdQuickGridDiscretization(const Teuchos::
   int* neighborhood = decomp.neighborhood.get();
   int neighborhoodIndex = 0;
   int numPointsWithZeroNeighbors = 0;
-  for(int i=0 ; i<decomp.numPoints ; ++i){
+  for(size_t i=0 ; i<decomp.numPoints ; ++i){
     int numNeighbors = neighborhood[neighborhoodIndex];
     if(numNeighbors > 0){
       numMyElements++;
@@ -145,7 +145,7 @@ PeridigmNS::PdQuickGridDiscretization::PdQuickGridDiscretization(const Teuchos::
   int* neighborhood = decomp->neighborhood.get();
   int neighborhoodIndex = 0;
   int numPointsWithZeroNeighbors = 0;
-  for(int i=0 ; i<decomp->numPoints ; ++i){
+  for(size_t i=0 ; i<decomp->numPoints ; ++i){
     int numNeighbors = neighborhood[neighborhoodIndex];
     if(numNeighbors > 0){
       numMyElements++;
